rename packages and write documentation in README.md
