+++ logprep/event_generator/http/manipulator.py
@@ -9,7 +9,7 @@ from functools import reduce
 from typing import TYPE_CHECKING, List
 
 if TYPE_CHECKING:
-    from logprep.load_generator.http.input import EventClassConfig
+    from logprep.event_generator.http.input import EventClassConfig
 
 
 class Manipulator:
