+++ logprep/event_generator/kafka/configuration.py
