+++ logprep/event_generator/kafka/document_loader.py
@@ -5,8 +5,8 @@ from logging import Logger
 
 import ndjson
 
-from logprep.load_generator.kafka.configuration import Configuration
-from logprep.load_generator.kafka.kafka_connector import KafkaConsumer
+from logprep.event_generator.kafka.configuration import Configuration
+from logprep.event_generator.kafka.kafka_connector import KafkaConsumer
 
 
 class DocumentLoader:
