+++ logprep/event_generator/kafka/kafka_connector.py
@@ -4,7 +4,7 @@ from typing import Optional
 
 from confluent_kafka import Consumer, Producer
 
-from logprep.load_generator.kafka.configuration import Kafka
+from logprep.event_generator.kafka.configuration import Kafka
 
 
 class KafkaProducer:
