+++ logprep/event_generator/kafka/util.py
@@ -4,7 +4,7 @@ from pathlib import Path
 
 import numpy as np
 
-from logprep.load_generator.kafka.configuration import Configuration
+from logprep.event_generator.kafka.configuration import Configuration
 
 
 def get_avg_size_mb(source_file: Path) -> float:
