+++ tests/testdata/event_generator/kafka/wineventlog_raw.jsonl
