+++ tests/unit/event_generator/http/test_controller.py
@@ -7,8 +7,8 @@ from unittest import mock
 
 import responses
 
-from logprep.load_generator.http.controller import Controller
-from tests.unit.load_generator.http.util import create_test_event_files
+from logprep.event_generator.http.controller import Controller
+from tests.unit.event_generator.http.util import create_test_event_files
 
 
 class TestController:
@@ -94,7 +94,7 @@ class TestController:
             assert call.request.headers.get("Content-Type") == expected_http_header
             assert call.response.status_code == 200
 
-    @mock.patch("logprep.load_generator.http.controller.ThreadPoolExecutor")
+    @mock.patch("logprep.event_generator.http.controller.ThreadPoolExecutor")
     def test_run_with_multiple_threads(self, mock_executor_class, tmp_path):
         self.contoller = Controller(
             input_dir="",
