+++ tests/unit/event_generator/http/test_input.py
@@ -10,8 +10,8 @@ import msgspec
 import pytest
 import yaml
 
-from logprep.load_generator.http.input import EventClassConfig, Input
-from tests.unit.load_generator.http.util import create_test_event_files
+from logprep.event_generator.http.input import EventClassConfig, Input
+from tests.unit.event_generator.http.util import create_test_event_files
 
 
 class TestInput:
