+++ tests/unit/event_generator/kafka/test_document_sender.py
@@ -8,9 +8,10 @@ import json
 from unittest import mock
 from unittest.mock import MagicMock, patch
 
-from logprep.load_generator.kafka.document_sender import DocumentSender
-from logprep.load_generator.kafka.kafka_connector import KafkaProducer
-from tests.testdata.load_generator.kafka.kafka_config_dict import get_config
+from tests.testdata.event_generator.kafka.kafka_config_dict import get_config
+
+from logprep.event_generator.kafka.document_sender import DocumentSender
+from logprep.event_generator.kafka.kafka_connector import KafkaProducer
 
 
 class MockedProducer:
@@ -32,14 +33,14 @@ class TestDocumentSender:
         mocked_kafka = MagicMock()
         mocked_kafka.Producer = MockedProducer
         with patch(
-            "logprep.load_generator.kafka.kafka_connector.Producer", return_value=mocked_kafka
+            "logprep.event_generator.kafka.kafka_connector.Producer", return_value=mocked_kafka
         ):
             self._document_sender = DocumentSender(get_config(), MagicMock())
 
     def test_init(self):
         assert isinstance(self._document_sender._kafka_producer, KafkaProducer)
 
-    @mock.patch("logprep.load_generator.kafka.document_sender.perf_counter")
+    @mock.patch("logprep.event_generator.kafka.document_sender.perf_counter")
     def test_send_zero(self, time_mock):
         time_mock.return_value = 0
         self._document_sender._kafka_producer._producer = MockedProducer()
@@ -49,7 +50,7 @@ class TestDocumentSender:
         assert count == 0
         assert self._document_sender._kafka_producer._producer.produced == []
 
-    @mock.patch("logprep.load_generator.kafka.document_sender.perf_counter")
+    @mock.patch("logprep.event_generator.kafka.document_sender.perf_counter")
     def test_send_one_of_zero(self, time_mock):
         time_mock.return_value = 0
         self._document_sender._kafka_producer._producer = MockedProducer()
@@ -57,7 +58,7 @@ class TestDocumentSender:
         assert count == 0
         assert self._document_sender._kafka_producer._producer.produced == []
 
-    @mock.patch("logprep.load_generator.kafka.document_sender.perf_counter")
+    @mock.patch("logprep.event_generator.kafka.document_sender.perf_counter")
     def test_send_one(self, time_mock):
         time_mock.return_value = 0
         self._document_sender._kafka_producer._producer = MockedProducer()
@@ -68,7 +69,7 @@ class TestDocumentSender:
         self._assert_uuid4_was_added_and_remove_it()
         assert self._document_sender._kafka_producer._producer.produced == [{"foo": "1"}]
 
-    @mock.patch("logprep.load_generator.kafka.document_sender.perf_counter")
+    @mock.patch("logprep.event_generator.kafka.document_sender.perf_counter")
     def test_send_multiple(self, time_mock):
         time_mock.return_value = 0
         self._document_sender._kafka_producer._producer = MockedProducer()
@@ -82,7 +83,7 @@ class TestDocumentSender:
             {"bar": "2"},
         ]
 
-    @mock.patch("logprep.load_generator.kafka.document_sender.perf_counter")
+    @mock.patch("logprep.event_generator.kafka.document_sender.perf_counter")
     def test_send_more_than_available(self, time_mock):
         time_mock.return_value = 0
         self._document_sender._kafka_producer._producer = MockedProducer()
