+++ tests/unit/event_generator/kafka/test_util.py
@@ -5,8 +5,8 @@
 from pathlib import Path
 from unittest.mock import MagicMock
 
-from logprep.load_generator.kafka.configuration import load_config
-from logprep.load_generator.kafka.util import (
+from logprep.event_generator.kafka.configuration import load_config
+from logprep.event_generator.kafka.util import (
     get_avg_size_mb,
     print_results,
     print_startup_info,
@@ -16,14 +16,14 @@ from logprep.load_generator.kafka.util import (
 class TestUtil:
     def test_get_avg_size_mb(self):
         assert (
-            get_avg_size_mb(Path("tests/testdata/load_generator/kafka/wineventlog_raw.jsonl"))
+            get_avg_size_mb(Path("tests/testdata/event_generator/kafka/wineventlog_raw.jsonl"))
             == 0.001621
         )
 
     def test_print_results_with_empty_shared_dict(self):
         config = load_config(
-            Path("tests/testdata/load_generator/kafka/config.yml"),
-            Path("tests/testdata/load_generator/kafka/wineventlog_raw.jsonl"),
+            Path("tests/testdata/event_generator/kafka/config.yml"),
+            Path("tests/testdata/event_generator/kafka/wineventlog_raw.jsonl"),
         )
         logger = MagicMock()
         print_results(config, logger, {})
@@ -31,8 +31,8 @@ class TestUtil:
 
     def test_print_results_with_filled_shared_dict(self):
         config = load_config(
-            Path("tests/testdata/load_generator/kafka/config.yml"),
-            Path("tests/testdata/load_generator/kafka/wineventlog_raw.jsonl"),
+            Path("tests/testdata/event_generator/kafka/config.yml"),
+            Path("tests/testdata/event_generator/kafka/wineventlog_raw.jsonl"),
         )
         logger = MagicMock()
         print_results(config, logger, {"foo_sent": 2, "bar_sent": 3, "foo_time": 4, "bar_time": 8})
@@ -40,8 +40,8 @@ class TestUtil:
 
     def test_print_startup_info_with_source_file(self):
         config = load_config(
-            Path("tests/testdata/load_generator/kafka/config.yml"),
-            Path("tests/testdata/load_generator/kafka/wineventlog_raw.jsonl"),
+            Path("tests/testdata/event_generator/kafka/config.yml"),
+            Path("tests/testdata/event_generator/kafka/wineventlog_raw.jsonl"),
         )
         logger = MagicMock()
         print_startup_info(config, logger)
@@ -49,8 +49,8 @@ class TestUtil:
 
     def test_print_startup_info_without_source_file(self):
         config = load_config(
-            Path("tests/testdata/load_generator/kafka/config.yml"),
-            Path("tests/testdata/load_generator/kafka/wineventlog_raw.jsonl"),
+            Path("tests/testdata/event_generator/kafka/config.yml"),
+            Path("tests/testdata/event_generator/kafka/wineventlog_raw.jsonl"),
         )
         config.source_file = None
         logger = MagicMock()
